Correciones realizadas, validaciones agregadas, ReadMe.txt mejorado, comentarios, controllers y funciones sin uso borrados.

+++ src/FibonacciWebApi/Services/IFibonacciService.cs
@@ -2,12 +2,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Numerics;
 
 namespace FibonacciWebApi.Services
 {
     public interface IFibonacciService
     {
-        long CalcularFibonacci(long i);
+        BigInteger CalcularFibonacci(int i);
         
     }
 }
